refactor(gui): delete dead Redux gui/state/ module, unify on foundation.state
gui/state/ (store.py/actions.py/selectors.py/signals.py/reducers/,
~3,700 lines) documented itself as "the" GUI state architecture, but
traced every real call site and found it wasn't: the fit/transform
wizard has run on foundation/state.py's plain mutable dataclasses all
along, and PipelineExecutionService's only use of gui.state (dispatching
pipeline_actions.* during execute_all()/execute_single_step()) wrote
into a global StateStore singleton that nothing in the running app ever
read back -- window.py's pipeline UI is wired directly to the service's
own Qt signals (step_started/step_completed/...), not to the store.
Two disconnected PipelineState/PipelineStepConfig representations
existed under the same names; only foundation.state's was ever read.

Moves the 6 dataclasses/enums genuinely used by 11 production files
(ParameterState, DatasetState, FitResult, BayesianResult, StepStatus,
PipelineStep) into foundation/state.py, drops the unused .clone()
method each carried (zero production callers), and deletes gui/state/
and gui/utils/pipeline_serializer.py (489 lines, zero callers anywhere)
along with the write-only _dispatch_action/_invoke_action/_run_action
plumbing in pipeline_execution_service.py.

Test triage was per-file, not blanket: test_visual_pipeline_state.py
and test_pipeline_serializer.py were 100% coupled to the deleted
machinery and are removed entirely; test_pipeline_step_config.py,
test_gui_integration.py, and test_gui_performance.py had real,
unrelated coverage (TransformService, widget/service integration,
non-state benchmarks) mixed in with the dead-code tests and were edited
class-by-class to keep the former and drop the latter.
test_pipeline_execution_service.py's status/running assertions read
state back through the deleted store; re-pointed to spy on the
service's own signals instead, which is the path production actually
uses -- net gain in coverage, not a loss.
