--- rheojax/gui/state/actions.py
@@ -1,1059 +1,0 @@
-"""State action creators for RheoJAX GUI.
-
-This module provides effector-style action functions that modify state through
-the StateStore and emit appropriate signals for UI reactivity.
-
-All action functions dispatch internally and return None.  Callers should NOT
-wrap them in ``store.dispatch()``."""
-
-import uuid
-from dataclasses import replace
-from datetime import datetime
-from pathlib import Path
-from typing import Any
-
-from rheojax.gui.state.store import (
-    AppState,
-    BayesianResult,
-    DatasetState,
-    FitResult,
-    ParameterState,
-    PipelineStep,
-    PipelineStepConfig,
-    StateStore,
-    StepStatus,
-    TransformRecord,
-    VisualPipelineState,
-)
-from rheojax.logging import get_logger
-
-logger = get_logger(__name__)
-
-# All action creators use the effector pattern: they dispatch internally and
-# return None.  Callers should NOT wrap them in store.dispatch().
-# R12-C-002 fix: effector actions that need atomicity use
-# store._atomic_state_and_signal() to wrap update_state + emit_signal.
-
-# Dataset Actions
-
-
-def load_dataset(
-    file_path: Path,
-    name: str,
-    test_mode: str,
-    x_data: Any,
-    y_data: Any,
-    y2_data: Any | None = None,
-    metadata: dict | None = None,
-) -> str:
-    """Load a dataset into the state.
-
-    Parameters
-    ----------
-    file_path : Path
-        Source file path
-    name : str
-        Dataset name
-    test_mode : str
-        Test mode (oscillation, relaxation, creep, rotation)
-    x_data : Any
-        X-axis data (JAX array)
-    y_data : Any
-        Y-axis data (JAX array)
-    y2_data : Any, optional
-        Second Y-axis data for oscillation (G'')
-    metadata : dict, optional
-        Additional metadata
-
-    Returns
-    -------
-    str
-        Dataset ID
-    """
-    store = StateStore()
-    dataset_id = str(uuid.uuid4())
-
-    logger.debug(
-        "load_dataset called",
-        name=name,
-        test_mode=test_mode,
-        file_path=str(file_path),
-    )
-
-    dataset = DatasetState(
-        id=dataset_id,
-        name=name,
-        file_path=file_path,
-        test_mode=test_mode,
-        x_data=x_data,
-        y_data=y_data,
-        y2_data=y2_data,
-        metadata=metadata or {},
-        is_modified=False,
-    )
-
-    def updater(state: AppState) -> AppState:
-        new_datasets = state.datasets.copy()
-        new_datasets[dataset_id] = dataset
-        return replace(
-            state,
-            datasets=new_datasets,
-            active_dataset_id=dataset_id,
-            is_modified=True,
-        )
-
-    # R12-C-002: use _atomic_state_and_signal for the primary signal;
-    # the secondary dataset_selected signal is emitted separately.
-    store._atomic_state_and_signal(updater, "dataset_added", dataset_id)
-    store.emit_signal("dataset_selected", dataset_id)
-
-    n_points = len(x_data) if hasattr(x_data, "__len__") else 0
-    logger.info(
-        "Dataset loaded",
-        dataset_id=dataset_id,
-        name=name,
-        n_points=n_points,
-        test_mode=test_mode,
-    )
-
-    return dataset_id
-
-
-def remove_dataset(dataset_id: str) -> None:
-    """Remove a dataset from state.
-
-    Parameters
-    ----------
-    dataset_id : str
-        Dataset identifier
-    """
-    store = StateStore()
-    logger.debug("remove_dataset called", dataset_id=dataset_id)
-
-    def updater(state: AppState) -> AppState:
-        new_datasets = state.datasets.copy()
-        if dataset_id in new_datasets:
-            del new_datasets[dataset_id]
-
-        # Clear active if removed
-        new_active = state.active_dataset_id
-        if new_active == dataset_id:
-            new_active = next(iter(new_datasets.keys()), None)
-
-        # Remove associated fit results
-        new_fit_results = {
-            k: v for k, v in state.fit_results.items() if v.dataset_id != dataset_id
-        }
-        new_bayesian_results = {
-            k: v
-            for k, v in state.bayesian_results.items()
-            if v.dataset_id != dataset_id
-        }
-
-        # Rescope Fit/Bayesian chips to the new active dataset -- a chip left
-        # COMPLETE from the just-removed dataset would otherwise keep showing
-        # green for a dataset that has no result at all.
-        pipeline = state.pipeline_state.clone()
-        pipeline.sync_fit_chip(
-            PipelineStep.FIT, new_fit_results, state.active_model_name, new_active
-        )
-        pipeline.sync_fit_chip(
-            PipelineStep.BAYESIAN,
-            new_bayesian_results,
-            state.active_model_name,
-            new_active,
-        )
-
-        return replace(
-            state,
-            datasets=new_datasets,
-            active_dataset_id=new_active,
-            fit_results=new_fit_results,
-            bayesian_results=new_bayesian_results,
-            pipeline_state=pipeline,
-            is_modified=True,
-        )
-
-    store._atomic_state_and_signal(updater, "dataset_removed", dataset_id)
-    logger.info("Dataset removed", dataset_id=dataset_id)
-
-
-def set_active_dataset(dataset_id: str) -> None:
-    """Set the active dataset.
-
-    Parameters
-    ----------
-    dataset_id : str
-        Dataset identifier
-    """
-    store = StateStore()
-    logger.debug("set_active_dataset called", dataset_id=dataset_id)
-
-    def updater(state: AppState) -> AppState:
-        if dataset_id not in state.datasets:
-            logger.warning(
-                "set_active_dataset: dataset not found", dataset_id=dataset_id
-            )
-            return state
-        return replace(state, active_dataset_id=dataset_id)
-
-    store._atomic_state_and_signal(updater, "dataset_selected", dataset_id)
-
-
-def update_dataset(dataset_id: str, **updates) -> None:
-    """Update dataset properties.
-
-    Parameters
-    ----------
-    dataset_id : str
-        Dataset identifier
-    **updates : dict
-        Properties to update
-    """
-    store = StateStore()
-    logger.debug(
-        "update_dataset called", dataset_id=dataset_id, update_keys=list(updates.keys())
-    )
-
-    def updater(state: AppState) -> AppState:
-        if dataset_id not in state.datasets:
-            logger.warning("update_dataset: dataset not found", dataset_id=dataset_id)
-            return state
-
-        dataset = state.datasets[dataset_id]
-        updated_dataset = replace(dataset, **updates)
-
-        new_datasets = state.datasets.copy()
-        new_datasets[dataset_id] = updated_dataset
-
-        new_fit_results = {
-            key: result
-            for key, result in state.fit_results.items()
-            if result.dataset_id != dataset_id
-        }
-        new_bayesian_results = {
-            key: result
-            for key, result in state.bayesian_results.items()
-            if result.dataset_id != dataset_id
-        }
-
-        pipeline = state.pipeline_state.clone()
-        if dataset_id == state.active_dataset_id:
-            pipeline.sync_fit_chip(
-                PipelineStep.FIT, new_fit_results, state.active_model_name, dataset_id
-            )
-            pipeline.sync_fit_chip(
-                PipelineStep.BAYESIAN,
-                new_bayesian_results,
-                state.active_model_name,
-                dataset_id,
-            )
-
-        return replace(
-            state,
-            datasets=new_datasets,
-            fit_results=new_fit_results,
-            bayesian_results=new_bayesian_results,
-            pipeline_state=pipeline,
-            is_modified=True,
-        )
-
-    store._atomic_state_and_signal(updater, "dataset_updated", dataset_id)
-
-
-# Model Actions
-
-
-def select_model(model_name: str, parameters: dict[str, ParameterState]) -> None:
-    """Select a model and initialize parameters.
-
-    Routes through dispatch() so the SET_ACTIVE_MODEL reducer handles both
-    active_model_name and model_params in a single atomic update.
-
-    Parameters
-    ----------
-    model_name : str
-        Model name
-    parameters : dict[str, ParameterState]
-        Model parameters
-    """
-    store = StateStore()
-    logger.debug("select_model called", model_name=model_name, n_params=len(parameters))
-    store.dispatch(
-        "SET_ACTIVE_MODEL", {"model_name": model_name, "model_params": parameters}
-    )
-    logger.info("Model selected", model_name=model_name, n_params=len(parameters))
-
-
-def update_parameter(name: str, value: float) -> None:
-    """Update a model parameter value.
-
-    Parameters
-    ----------
-    name : str
-        Parameter name
-    value : float
-        New parameter value
-    """
-    store = StateStore()
-    captured_model_name = [None]
-    logger.debug("update_parameter called", name=name, value=value)
-
-    def updater(state: AppState) -> AppState:
-        captured_model_name[0] = state.active_model_name
-        if name not in state.model_params:
-            logger.warning("update_parameter: param not found", name=name)
-            return state
-
-        param = state.model_params[name]
-        updated_param = replace(param, value=value)
-
-        new_params = state.model_params.copy()
-        new_params[name] = updated_param
-
-        return replace(state, model_params=new_params)
-
-    # R12-C-002 note: signal args depend on state read inside the updater,
-    # so _atomic_state_and_signal cannot be used here.  The captured
-    # model_name is read under the store lock (inside update_state),
-    # making the signal arg semantically correct.
-    store.update_state(updater)
-    model_name = captured_model_name[0]
-
-    if model_name:
-        store.emit_signal("model_params_changed", model_name)
-
-
-def update_parameter_bounds(name: str, min_bound: float, max_bound: float) -> None:
-    """Update bounds for a model parameter.
-
-    Parameters
-    ----------
-    name : str
-        Parameter name
-    min_bound : float
-        New minimum bound
-    max_bound : float
-        New maximum bound
-    """
-    store = StateStore()
-    captured_model_name = [None]
-    logger.debug(
-        "update_parameter_bounds called",
-        name=name,
-        min_bound=min_bound,
-        max_bound=max_bound,
-    )
-
-    def updater(state: AppState) -> AppState:
-        captured_model_name[0] = state.active_model_name
-        if name not in state.model_params:
-            logger.warning("update_parameter_bounds: param not found", name=name)
-            return state
-
-        param = state.model_params[name]
-        updated_param = replace(
-            param, min_bound=float(min_bound), max_bound=float(max_bound)
-        )
-
-        new_params = state.model_params.copy()
-        new_params[name] = updated_param
-
-        return replace(state, model_params=new_params)
-
-    # R12-C-002 note: signal args depend on state read inside the updater.
-    store.update_state(updater)
-    model_name = captured_model_name[0]
-
-    if model_name:
-        store.emit_signal("model_params_changed", model_name)
-
-
-def toggle_parameter_fixed(name: str, fixed: bool) -> None:
-    """Toggle fixed state for a model parameter.
-
-    Parameters
-    ----------
-    name : str
-        Parameter name
-    fixed : bool
-        Whether the parameter is fixed
-    """
-    store = StateStore()
-    captured_model_name = [None]
-    logger.debug("toggle_parameter_fixed called", name=name, fixed=fixed)
-
-    def updater(state: AppState) -> AppState:
-        captured_model_name[0] = state.active_model_name
-        if name not in state.model_params:
-            logger.warning("toggle_parameter_fixed: param not found", name=name)
-            return state
-
-        param = state.model_params[name]
-        updated_param = replace(param, fixed=bool(fixed))
-
-        new_params = state.model_params.copy()
-        new_params[name] = updated_param
-
-        return replace(state, model_params=new_params)
-
-    # R12-C-002 note: signal args depend on state read inside the updater.
-    store.update_state(updater)
-    model_name = captured_model_name[0]
-
-    if model_name:
-        store.emit_signal("model_params_changed", model_name)
-
-
-def reset_parameters(default_params: dict[str, ParameterState]) -> None:
-    """Reset parameters to default values.
-
-    Parameters
-    ----------
-    default_params : dict[str, ParameterState]
-        Default parameter states
-    """
-    store = StateStore()
-    _model_name: list[str | None] = [None]
-    logger.debug("reset_parameters called", n_params=len(default_params))
-
-    def updater(state: AppState) -> AppState:
-        # R10-AC-002: Read model_name inside the updater closure (under the
-        # store lock) to eliminate the TOCTOU race between get_state() and
-        # update_state() where active_model_name could change in between.
-        _model_name[0] = state.active_model_name
-        return replace(state, model_params=default_params)
-
-    # R12-C-002 note: signal args depend on state read inside the updater.
-    store.update_state(updater)
-
-    if _model_name[0]:
-        logger.info(
-            "Parameters reset to defaults",
-            model_name=_model_name[0],
-            n_params=len(default_params),
-        )
-        store.emit_signal("model_params_changed", _model_name[0])
-
-
-# Fit Actions
-
-
-def store_fit_result(result: FitResult) -> None:
-    """Store a completed fit result.
-
-    Routes through dispatch() to ensure pipeline state is updated
-    consistently (STORE_FIT_RESULT reducer sets FIT step to COMPLETE).
-
-    Parameters
-    ----------
-    result : FitResult
-        Fit result
-    """
-    store = StateStore()
-    model_name = getattr(result, "model_name", "unknown")
-    dataset_id = getattr(result, "dataset_id", "unknown")
-    logger.debug(
-        "store_fit_result called",
-        model_name=model_name,
-        dataset_id=dataset_id,
-    )
-    store.dispatch(
-        "STORE_FIT_RESULT",
-        {
-            "model_name": model_name,
-            "dataset_id": dataset_id,
-            "result": result,
-        },
-    )
-    logger.info(
-        "Fit result stored",
-        model_name=model_name,
-        dataset_id=dataset_id,
-        success=getattr(result, "success", None),
-        r_squared=getattr(result, "r_squared", None),
-    )
-
-
-def set_active_model(model_name: str) -> None:
-    """Set the active model (dispatches internally).
-
-    Parameters
-    ----------
-    model_name : str
-        Model name
-    """
-    store = StateStore()
-    logger.debug("set_active_model called", model_name=model_name)
-    store.dispatch("SET_ACTIVE_MODEL", {"model_name": model_name})
-
-
-def start_fitting(model_name: str, dataset_id: str) -> None:
-    """Mark fitting as started (dispatches internally).
-
-    Parameters
-    ----------
-    model_name : str
-        Model name
-    dataset_id : str
-        Dataset identifier
-    """
-    store = StateStore()
-    logger.debug("start_fitting called", model_name=model_name, dataset_id=dataset_id)
-    store.dispatch(
-        "START_FITTING", {"model_name": model_name, "dataset_id": dataset_id}
-    )
-
-
-def fitting_completed(result: FitResult) -> None:
-    """Mark fit as completed (dispatches internally).
-
-    Parameters
-    ----------
-    result : FitResult
-        Fit result
-    """
-    store = StateStore()
-    logger.debug("fitting_completed called")
-    store.dispatch("FITTING_COMPLETED", {"result": result})
-
-
-def fitting_failed(error: str) -> None:
-    """Mark fit as failed (dispatches internally).
-
-    Parameters
-    ----------
-    error : str
-        Error message
-    """
-    store = StateStore()
-    logger.debug("fitting_failed called", error=error)
-    store.dispatch("FITTING_FAILED", {"error": error})
-
-
-# Bayesian Actions
-
-
-def start_bayesian(model_name: str, dataset_id: str) -> None:
-    """Mark Bayesian inference as started (dispatches internally).
-
-    Parameters
-    ----------
-    model_name : str
-        Model name
-    dataset_id : str
-        Dataset identifier
-    """
-    store = StateStore()
-    logger.debug("start_bayesian called", model_name=model_name, dataset_id=dataset_id)
-    store.dispatch(
-        "START_BAYESIAN", {"model_name": model_name, "dataset_id": dataset_id}
-    )
-
-
-def store_bayesian_result(result: BayesianResult) -> None:
-    """Store a completed Bayesian result.
-
-    Routes through dispatch() to ensure pipeline state is updated
-    consistently (STORE_BAYESIAN_RESULT reducer sets BAYESIAN step to COMPLETE).
-
-    Parameters
-    ----------
-    result : BayesianResult
-        Bayesian result
-    """
-    store = StateStore()
-    model_name = getattr(result, "model_name", "unknown")
-    dataset_id = getattr(result, "dataset_id", "unknown")
-    logger.debug(
-        "store_bayesian_result called",
-        model_name=model_name,
-        dataset_id=dataset_id,
-    )
-    store.dispatch(
-        "STORE_BAYESIAN_RESULT",
-        {
-            "model_name": model_name,
-            "dataset_id": dataset_id,
-            "result": result,
-        },
-    )
-    logger.info(
-        "Bayesian result stored",
-        model_name=model_name,
-        dataset_id=dataset_id,
-        divergences=getattr(result, "divergences", None),
-        mcmc_time=getattr(result, "mcmc_time", None),
-    )
-
-
-def update_bayesian_progress(progress: float) -> None:
-    """Update Bayesian inference progress (dispatches internally).
-
-    Parameters
-    ----------
-    progress : float
-        Progress percentage (0-100)
-    """
-    store = StateStore()
-    store.dispatch("BAYESIAN_PROGRESS", {"progress": progress})
-
-
-def bayesian_completed(result: BayesianResult) -> None:
-    """Mark Bayesian inference as completed (dispatches internally).
-
-    Parameters
-    ----------
-    result : BayesianResult
-        Bayesian inference result
-    """
-    store = StateStore()
-    logger.debug("bayesian_completed called")
-    store.dispatch("BAYESIAN_COMPLETED", {"result": result})
-
-
-def bayesian_failed(error: str) -> None:
-    """Mark Bayesian inference as failed (dispatches internally).
-
-    Parameters
-    ----------
-    error : str
-        Error message
-    """
-    store = StateStore()
-    logger.debug("bayesian_failed called", error=error)
-    store.dispatch("BAYESIAN_FAILED", {"error": error})
-
-
-# Pipeline Actions
-
-
-def set_pipeline_step(step: PipelineStep, status: StepStatus) -> None:
-    """Update pipeline step status.
-
-    Routes through dispatch() for consistent signal emission.
-
-    Parameters
-    ----------
-    step : PipelineStep
-        Pipeline step
-    status : StepStatus
-        New status
-    """
-    store = StateStore()
-    logger.debug("set_pipeline_step called", step=step.name, status=status.name)
-    store.dispatch(
-        "SET_PIPELINE_STEP",
-        {"step": step.name, "status": status.name},
-    )
-
-
-# JAX Actions
-
-
-def set_jax_device(device: str) -> None:
-    """Set JAX device.
-
-    Parameters
-    ----------
-    device : str
-        Device name (cpu, cuda, tpu)
-    """
-    store = StateStore()
-    logger.info("JAX device set", device=device)
-
-    def updater(state: AppState) -> AppState:
-        return replace(state, jax_device=device)
-
-    store._atomic_state_and_signal(updater, "jax_device_changed", device)
-
-
-# Settings Actions
-
-
-def set_theme(theme: str) -> None:
-    """Set UI theme.
-
-    Parameters
-    ----------
-    theme : str
-        Theme name (light, dark)
-    """
-    store = StateStore()
-    logger.info("Theme changed", theme=theme)
-
-    def updater(state: AppState) -> AppState:
-        return replace(state, theme=theme)
-
-    store._atomic_state_and_signal(updater, "theme_changed", theme, track_undo=False)
-
-
-def set_seed(seed: int) -> None:
-    """Set random seed.
-
-    Parameters
-    ----------
-    seed : int
-        Random seed
-    """
-    store = StateStore()
-    logger.debug("set_seed called", seed=seed)
-
-    def updater(state: AppState) -> AppState:
-        return replace(state, current_seed=seed)
-
-    store.update_state(updater, track_undo=False)
-
-
-# Project Actions
-
-
-def save_project(path: Path) -> None:
-    """Mark project as saved.
-
-    Parameters
-    ----------
-    path : Path
-        Project file path
-    """
-    store = StateStore()
-    logger.debug("save_project called", path=str(path))
-
-    def updater(state: AppState) -> AppState:
-        # Add to recent projects
-        recent = [p for p in state.recent_projects if p != path]
-        recent.insert(0, path)
-        recent = recent[:10]  # Keep last 10
-
-        return replace(
-            state,
-            project_path=path,
-            project_name=path.stem,
-            is_modified=False,
-            recent_projects=recent,
-        )
-
-    store._atomic_state_and_signal(
-        updater, "project_saved", str(path), track_undo=False
-    )
-    logger.info("Project saved", path=str(path))
-
-
-def load_project(path: Path, state: AppState) -> None:
-    """Load project from file.
-
-    Parameters
-    ----------
-    path : Path
-        Project file path
-    state : AppState
-        Loaded application state (used as the base for the new store state)
-    """
-    store = StateStore()
-    logger.debug(
-        "load_project called",
-        path=str(path),
-        n_datasets=len(state.datasets),
-    )
-
-    # R12-C-012: Use an updater function that reads current_state from the
-    # store rather than capturing the caller-supplied snapshot in a lambda.
-    # This prevents a TOCTOU race where a concurrent dispatch modifies
-    # recent_projects between load_project() being called and update_state()
-    # applying the change — the updater always works from the latest state.
-    def _load_project_updater(current_state: AppState) -> AppState:
-        recent = [p for p in current_state.recent_projects if p != path]
-        recent.insert(0, path)
-        recent = recent[:10]
-        return replace(
-            state,  # start from the loaded snapshot (full project state)
-            project_path=path,
-            project_name=path.stem,
-            is_modified=False,
-            recent_projects=recent,
-        )
-
-    # R12-C-002 note: cannot use _atomic_state_and_signal because
-    # clear_history() must be called between update and signal.
-    store.update_state(_load_project_updater, track_undo=False)
-    store.clear_history()
-
-    logger.info(
-        "Project loaded",
-        path=str(path),
-        n_datasets=len(state.datasets),
-    )
-    store.emit_signal("project_loaded", str(path))
-
-
-def add_transform_record(
-    source_id: str,
-    target_id: str,
-    transform_name: str,
-    parameters: dict[str, Any],
-    seed: int | None = None,
-) -> None:
-    """Add a transform record for provenance tracking.
-
-    Parameters
-    ----------
-    source_id : str
-        Source dataset ID
-    target_id : str
-        Target dataset ID
-    transform_name : str
-        Transform name
-    parameters : dict[str, Any]
-        Transform parameters
-    seed : int, optional
-        Random seed used
-    """
-    store = StateStore()
-    logger.debug(
-        "add_transform_record called",
-        transform_name=transform_name,
-        source_id=source_id,
-        target_id=target_id,
-        seed=seed,
-    )
-
-    record = TransformRecord(
-        timestamp=datetime.now(),
-        source_dataset_id=source_id,
-        target_dataset_id=target_id,
-        transform_name=transform_name,
-        parameters=parameters,
-        seed=seed,
-    )
-
-    def updater(state: AppState) -> AppState:
-        new_history = state.transform_history.copy()
-        new_history.append(record)
-        return replace(state, transform_history=new_history)
-
-    store._atomic_state_and_signal(
-        updater, "transform_applied", transform_name, target_id
-    )
-    logger.info(
-        "Transform record added",
-        transform_name=transform_name,
-        source_id=source_id,
-        target_id=target_id,
-    )
-
-
-# --- Visual Pipeline Actions ---
-
-
-def add_pipeline_step(
-    step_type: str,
-    name: str,
-    config: dict[str, Any] | None = None,
-    position: int | None = None,
-) -> str:
-    """Add a step to the visual pipeline.
-
-    Parameters
-    ----------
-    step_type : str
-        Step type: "load", "transform", "fit", "bayesian", "export"
-    name : str
-        Display name for the step
-    config : dict, optional
-        Step configuration
-    position : int, optional
-        Desired insertion position (ignored — steps are appended and
-        reindexed by the reducer)
-
-    Returns
-    -------
-    str
-        New step UUID
-    """
-    step_id = str(uuid.uuid4())
-    step_config = PipelineStepConfig(
-        id=step_id,
-        step_type=step_type,
-        name=name,
-        config=config or {},
-        status=StepStatus.PENDING,
-        position=position or 0,
-    )
-    store = StateStore()
-    logger.debug(
-        "add_pipeline_step called",
-        step_id=step_id,
-        step_type=step_type,
-        name=name,
-    )
-    store.dispatch("ADD_PIPELINE_STEP", {"step_config": step_config})
-    return step_id
-
-
-def remove_pipeline_step(step_id: str) -> None:
-    """Remove a step from the visual pipeline.
-
-    Parameters
-    ----------
-    step_id : str
-        UUID of the step to remove
-    """
-    store = StateStore()
-    logger.debug("remove_pipeline_step called", step_id=step_id)
-    store.dispatch("REMOVE_PIPELINE_STEP", {"step_id": step_id})
-
-
-def reorder_pipeline_step(step_id: str, new_position: int) -> None:
-    """Move a pipeline step to a new position.
-
-    Parameters
-    ----------
-    step_id : str
-        UUID of the step to move
-    new_position : int
-        Target zero-based index
-    """
-    store = StateStore()
-    logger.debug(
-        "reorder_pipeline_step called",
-        step_id=step_id,
-        new_position=new_position,
-    )
-    store.dispatch(
-        "REORDER_PIPELINE_STEP",
-        {"step_id": step_id, "new_position": new_position},
-    )
-
-
-def select_pipeline_step(step_id: str | None) -> None:
-    """Select a pipeline step (or deselect with None).
-
-    Parameters
-    ----------
-    step_id : str | None
-        UUID to select, or None to clear selection
-    """
-    store = StateStore()
-    logger.debug("select_pipeline_step called", step_id=step_id)
-    store.dispatch("SELECT_PIPELINE_STEP", {"step_id": step_id})
-
-
-def update_step_config(step_id: str, config: dict[str, Any]) -> None:
-    """Update the configuration of a pipeline step.
-
-    Parameters
-    ----------
-    step_id : str
-        UUID of the target step
-    config : dict
-        Key/value pairs to merge into the step's config
-    """
-    store = StateStore()
-    logger.debug(
-        "update_step_config called",
-        step_id=step_id,
-        config_keys=list(config.keys()),
-    )
-    store.dispatch("UPDATE_STEP_CONFIG", {"step_id": step_id, "config": config})
-
-
-def update_step_status(
-    step_id: str,
-    status: StepStatus,
-    error_message: str | None = None,
-) -> None:
-    """Update the execution status of a pipeline step.
-
-    Parameters
-    ----------
-    step_id : str
-        UUID of the target step
-    status : StepStatus
-        New status value
-    error_message : str, optional
-        Error description (used when status is ERROR)
-    """
-    store = StateStore()
-    logger.debug(
-        "update_step_status called",
-        step_id=step_id,
-        status=status.name,
-    )
-    store.dispatch(
-        "UPDATE_STEP_STATUS",
-        {"step_id": step_id, "status": status, "error_message": error_message},
-    )
-
-
-def cache_step_result(step_id: str, result: Any) -> None:
-    """Cache the result of a completed pipeline step.
-
-    Parameters
-    ----------
-    step_id : str
-        UUID of the completed step
-    result : Any
-        Result object to cache
-    """
-    store = StateStore()
-    logger.debug("cache_step_result called", step_id=step_id)
-    store.dispatch("CACHE_STEP_RESULT", {"step_id": step_id, "result": result})
-
-
-def set_pipeline_running(is_running: bool, current_step_id: str | None = None) -> None:
-    """Set pipeline execution state.
-
-    Parameters
-    ----------
-    is_running : bool
-        True when the pipeline is actively executing
-    current_step_id : str, optional
-        UUID of the step currently executing
-    """
-    store = StateStore()
-    logger.debug(
-        "set_pipeline_running called",
-        is_running=is_running,
-        current_step_id=current_step_id,
-    )
-    store.dispatch(
-        "SET_PIPELINE_RUNNING",
-        {"is_running": is_running, "current_step_id": current_step_id},
-    )
-
-
-def set_pipeline_name(name: str) -> None:
-    """Set the visual pipeline name.
-
-    Parameters
-    ----------
-    name : str
-        New pipeline name
-    """
-    store = StateStore()
-    logger.debug("set_pipeline_name called", name=name)
-    store.dispatch("SET_PIPELINE_NAME", {"name": name})
-
-
-def clear_pipeline() -> None:
-    """Clear the entire visual pipeline (steps, results, selection)."""
-    store = StateStore()
-    logger.debug("clear_pipeline called")
-    store.dispatch("CLEAR_PIPELINE", {})
-
-
-def load_pipeline(visual_pipeline: VisualPipelineState) -> None:
-    """Load a visual pipeline state (e.g., restored from YAML/project file).
-
-    Parameters
-    ----------
-    visual_pipeline : VisualPipelineState
-        Pipeline state to load
-    """
-    store = StateStore()
-    logger.debug(
-        "load_pipeline called",
-        pipeline_name=visual_pipeline.pipeline_name,
-        n_steps=len(visual_pipeline.steps),
-    )
-    store.dispatch("LOAD_PIPELINE", {"visual_pipeline": visual_pipeline})
