--- rheojax/gui/state/reducers/ui_reducers.py
@@ -1,130 +1,0 @@
-"""UI and settings action reducers.
-
-Handles: SET_THEME, SET_OS_THEME, SET_WORKFLOW_MODE, SET_TAB, NAVIGATE_TAB,
-UPDATE_PREFERENCES, CHECK_COMPATIBILITY.
-"""
-
-from __future__ import annotations
-
-from collections.abc import Callable
-from dataclasses import replace
-from typing import TYPE_CHECKING, Any
-
-if TYPE_CHECKING:
-    from rheojax.gui.state.store import AppState, StateStore
-
-from rheojax.logging import get_logger
-
-logger = get_logger(__name__)
-
-
-def reduce_set_theme(
-    action: dict[str, Any],
-) -> Callable[[AppState], AppState]:
-    theme = action.get("theme", "light")
-
-    def updater(state: AppState) -> AppState:
-        if state.theme == theme:
-            return state
-        return replace(state, theme=theme, is_modified=True)
-
-    return updater
-
-
-def reduce_set_os_theme(
-    action: dict[str, Any],
-) -> Callable[[AppState], AppState]:
-    os_theme = action.get("os_theme", "light")
-
-    def updater(state: AppState) -> AppState:
-        if state.os_theme == os_theme:
-            return state
-        return replace(state, os_theme=os_theme)
-
-    return updater
-
-
-def reduce_set_workflow_mode(
-    action: dict[str, Any],
-) -> Callable[[AppState], AppState]:
-    from rheojax.gui.state.store import WorkflowMode
-
-    mode = action.get("mode")
-
-    def updater(state: AppState) -> AppState:
-        new_mode = WorkflowMode[mode.upper()] if isinstance(mode, str) else mode
-        if isinstance(new_mode, WorkflowMode):
-            return replace(state, workflow_mode=new_mode)
-        return state
-
-    return updater
-
-
-def reduce_set_tab(
-    action: dict[str, Any],
-) -> Callable[[AppState], AppState]:
-    tab = action.get("tab")
-
-    def updater(state: AppState) -> AppState:
-        resolved_tab = tab if tab is not None else state.current_tab
-        return replace(state, current_tab=resolved_tab)
-
-    return updater
-
-
-def reduce_update_preferences(
-    action: dict[str, Any],
-    store: StateStore,
-) -> Callable[[AppState], AppState]:
-    prefs = action
-
-    def updater(state: AppState) -> AppState:
-        updates: dict[str, Any] = {}
-        for key in (
-            "theme",
-            "last_export_dir",
-            "current_seed",
-        ):
-            if key in prefs:
-                updates[key] = prefs[key]
-        # PreferencesDialog.get_preferences() emits "autosave_enabled";
-        # AppState's field is "auto_save_enabled" -- map explicitly rather
-        # than relying on the key names matching.
-        if "autosave_enabled" in prefs:
-            updates["auto_save_enabled"] = prefs["autosave_enabled"]
-        return replace(state, **updates) if updates else state
-
-    # Apply runtime settings that live outside AppState
-    if "max_undo_steps" in prefs:
-        store.set_max_undo_size(int(prefs["max_undo_steps"]))
-    if "worker_isolation_mode" in prefs:
-        import os
-
-        os.environ["RHEOJAX_WORKER_ISOLATION"] = prefs["worker_isolation_mode"]
-
-    handled_keys = {
-        "type",
-        "theme",
-        "last_export_dir",
-        "current_seed",
-        "autosave_enabled",
-        "max_undo_steps",
-        "worker_isolation_mode",
-    }
-    unhandled = set(prefs) - handled_keys
-    if unhandled:
-        # Keys PreferencesDialog exposes (device, plot_style, etc.) that this
-        # reducer does not yet apply -- log instead of dropping silently.
-        logger.warning("Unhandled preference keys", keys=sorted(unhandled))
-
-    return updater
-
-
-def reduce_check_compatibility(
-    action: dict[str, Any],
-) -> Callable[[AppState], AppState] | None:
-    # STORE-002: CHECK_COMPATIBILITY is a UI-only trigger (opens the
-    # diagnostics tab) and does not need to mutate state.  Return
-    # None so dispatch() skips update_state entirely (no undo entry,
-    # no state clone, no state_changed signal).
-    return None
