+++ tests/gui/test_pipeline_execution_service.py
@@ -1,13 +1,16 @@
 """Tests for the pipeline execution service.
 
-Unit tests that verify status dispatch, context accumulation, and error
+Unit tests that verify signal emission, context accumulation, and error
 handling in PipelineExecutionService.  All service calls are mocked so
 that no real data I/O, model fitting, or Qt event loop is needed.
 
 PipelineExecutionService subclasses QObject, so a QApplication must exist
 before instantiation.  All tests that touch the service are guarded by
-``pytest.mark.skipif(not HAS_PYSIDE6, ...)``.  State-only assertions use
-the real StateStore and reset it via the autouse fixture from conftest.py.
+``pytest.mark.skipif(not HAS_PYSIDE6, ...)``.  Step/pipeline status is
+signal-based in production (step_started/step_completed/step_failed/
+pipeline_started/pipeline_completed/pipeline_failed) -- window.py connects
+to these signals directly, so status-related assertions here spy on the
+same signals rather than reading a shared store.
 """
 
 import os
@@ -19,13 +22,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from rheojax.gui.state.actions import add_pipeline_step
-from rheojax.gui.state.selectors import (
-    get_pipeline_step_by_id,
-    get_pipeline_step_result,
-    is_pipeline_running,
-)
-from rheojax.gui.state.store import StateStore, StepStatus
+from rheojax.gui.foundation.state import PipelineStepConfig
 
 # ---------------------------------------------------------------------------
 # PySide6 availability guard
@@ -47,14 +44,6 @@ pytestmark = pytest.mark.skipif(
 # ---------------------------------------------------------------------------
 # Fixtures
 # ---------------------------------------------------------------------------
-
-
-@pytest.fixture(autouse=True)
-def reset_store():
-    """Reset the StateStore singleton before and after every test."""
-    StateStore.reset()
-    yield
-    StateStore.reset()
 
 
 @pytest.fixture(scope="module")
@@ -81,30 +70,25 @@ def service(qapp_module):
 # ---------------------------------------------------------------------------
 # Helper builders
 # ---------------------------------------------------------------------------
+# PipelineExecutionService only ever reads .id/.step_type/.config off a step
+# (confirmed against batch_runner.py's real construction) -- these mirror
+# that shape directly, no add_pipeline_step()/state-store round-trip needed.
 
 
-def _add_load_step() -> str:
-    return add_pipeline_step("load", "Load", config={"file": "/tmp/data.csv"})
-
-
-def _add_fit_step() -> str:
-    return add_pipeline_step("fit", "Fit", config={"model": "Maxwell"})
-
-
-def _add_transform_step() -> str:
-    return add_pipeline_step("transform", "FFT", config={"name": "fft"})
-
-
-def _add_bayesian_step() -> str:
-    return add_pipeline_step(
-        "bayesian",
-        "Bayes",
-        config={"num_warmup": 100, "num_samples": 200},
+def _load_step() -> PipelineStepConfig:
+    return PipelineStepConfig(
+        id="load-1", step_type="load", config={"file": "/tmp/data.csv"}
     )
 
 
-def _add_export_step() -> str:
-    return add_pipeline_step("export", "Export", config={"output": "/tmp/out/"})
+def _fit_step() -> PipelineStepConfig:
+    return PipelineStepConfig(id="fit-1", step_type="fit", config={"model": "Maxwell"})
+
+
+def _transform_step() -> PipelineStepConfig:
+    return PipelineStepConfig(
+        id="transform-1", step_type="transform", config={"name": "fft"}
+    )
 
 
 # ---------------------------------------------------------------------------
@@ -134,9 +118,10 @@ class TestPipelineExecutionService:
 
     @pytest.mark.smoke
     def test_execute_step_updates_status_to_active_then_complete(self, service):
-        """execute_single_step must set ACTIVE then COMPLETE on a successful step."""
-        step_id = _add_load_step()
-        step = get_pipeline_step_by_id(step_id)
+        """execute_single_step must emit step_started then step_completed on success."""
+        step = _load_step()
+        completed_ids = []
+        service.step_completed.connect(completed_ids.append)
 
         mock_data = MagicMock(name="RheoData")
         mock_data.metadata = {}
@@ -145,40 +130,25 @@ class TestPipelineExecutionService:
             context: dict = {}
             service.execute_single_step(step, context)
 
-        final_step = get_pipeline_step_by_id(step_id)
-        assert final_step.status == StepStatus.COMPLETE
+        assert completed_ids == [step.id]
 
     @pytest.mark.smoke
     def test_execute_step_caches_result(self, service):
-        """execute_single_step must cache the step result in state.
-
-        The state store deep-copies step_results on every clone, so the cached
-        value may not be the same object.  We verify the result is not None
-        (i.e., something was cached) rather than using identity.
-        """
-        step_id = _add_load_step()
-        step = get_pipeline_step_by_id(step_id)
+        """execute_single_step must return the step result to the caller."""
+        step = _load_step()
 
         mock_data = MagicMock(name="RheoData")
         mock_data.metadata = {}
 
         with patch.object(service, "_execute_load", return_value=mock_data):
             context: dict = {}
-            service.execute_single_step(step, context)
+            result = service.execute_single_step(step, context)
 
-        assert get_pipeline_step_result(step_id) is not None
+        assert result is mock_data
 
     def test_execute_step_unknown_type_raises(self, service):
         """_execute_step must raise ValueError for an unknown step type."""
-        from rheojax.gui.state.store import PipelineStepConfig
-
-        bad_step = PipelineStepConfig(
-            id="bad-id",
-            step_type="not_a_real_type",
-            name="Bad",
-            config={},
-            status=StepStatus.PENDING,
-        )
+        bad_step = PipelineStepConfig(id="bad-id", step_type="not_a_real_type", config={})
         with pytest.raises(ValueError, match="Unknown step type"):
             service._execute_step(bad_step, {})
 
@@ -229,9 +199,12 @@ class TestPipelineExecutionService:
             )
 
     def test_execute_single_step_marks_error_on_failed_fit(self, service):
-        """A failed fit step must land on StepStatus.ERROR, not COMPLETE."""
-        step_id = _add_fit_step()
-        step = get_pipeline_step_by_id(step_id)
+        """A failed fit step must emit step_failed, not step_completed."""
+        step = _fit_step()
+        failed_ids = []
+        completed_ids = []
+        service.step_failed.connect(lambda step_id, msg: failed_ids.append(step_id))
+        service.step_completed.connect(completed_ids.append)
 
         mock_data = MagicMock()
         failed_result = MagicMock(success=False, message="Fit diverged")
@@ -241,8 +214,8 @@ class TestPipelineExecutionService:
         with pytest.raises(RuntimeError, match="Fit diverged"):
             service.execute_single_step(step, {"data": mock_data})
 
-        final_step = get_pipeline_step_by_id(step_id)
-        assert final_step.status == StepStatus.ERROR
+        assert failed_ids == [step.id]
+        assert completed_ids == []
 
     def test_execute_bayesian_requires_data_in_context(self, service):
         """_execute_bayesian must raise ValueError when context has no 'data'."""
@@ -267,9 +240,10 @@ class TestPipelineExecutionService:
         assert len(completed_calls) == 1
 
     def test_execute_step_sets_error_status_on_failure(self, service):
-        """When a step handler raises, the step status must become ERROR."""
-        step_id = _add_load_step()
-        step = get_pipeline_step_by_id(step_id)
+        """When a step handler raises, the service must emit step_failed with the error."""
+        step = _load_step()
+        failed = []
+        service.step_failed.connect(lambda step_id, msg: failed.append((step_id, msg)))
 
         with patch.object(
             service,
@@ -280,24 +254,26 @@ class TestPipelineExecutionService:
             with pytest.raises(RuntimeError, match="disk error"):
                 service.execute_single_step(step, context)
 
-        failed_step = get_pipeline_step_by_id(step_id)
-        assert failed_step.status == StepStatus.ERROR
-        assert "disk error" in (failed_step.error_message or "")
+        assert len(failed) == 1
+        assert failed[0][0] == step.id
+        assert "disk error" in failed[0][1]
 
     def test_execute_all_sets_running_true_during_execution(self, service):
-        """execute_all must mark the pipeline as running before the first step."""
-        running_states: list[bool] = []
+        """execute_all must emit pipeline_started before the first step, and
+        pipeline_completed only after the last step finishes."""
+        events: list[str] = []
 
-        step_id = _add_load_step()
-        step = get_pipeline_step_by_id(step_id)
+        step = _load_step()
         mock_data = MagicMock()
         mock_data.metadata = {}
 
-        # Capture is_pipeline_running() at the moment the step starts.
+        service.pipeline_started.connect(lambda: events.append("started"))
+        service.pipeline_completed.connect(lambda: events.append("completed"))
+
         original_execute = service._execute_step
 
         def capturing_execute(s, ctx):
-            running_states.append(is_pipeline_running())
+            events.append("step")
             return original_execute(s, ctx)
 
         with (
@@ -306,11 +282,10 @@ class TestPipelineExecutionService:
         ):
             service.execute_all([step])
 
-        assert any(running_states), (
-            "is_pipeline_running() was never True during execution"
+        assert events == ["started", "step", "completed"], (
+            "pipeline_started must fire before the step runs, "
+            "pipeline_completed only after it finishes"
         )
-        # After completion the pipeline must be stopped.
-        assert is_pipeline_running() is False
 
     # ------------------------------------------------------------------
     # I7: Multi-step context accumulation
@@ -318,13 +293,9 @@ class TestPipelineExecutionService:
 
     def test_execute_all_passes_context_between_steps(self, service):
         """execute_all must accumulate context across steps: load → transform → fit."""
-        load_id = _add_load_step()
-        transform_id = _add_transform_step()
-        fit_id = _add_fit_step()
-
-        load_step = get_pipeline_step_by_id(load_id)
-        transform_step = get_pipeline_step_by_id(transform_id)
-        fit_step = get_pipeline_step_by_id(fit_id)
+        load_step = _load_step()
+        transform_step = _transform_step()
+        fit_step = _fit_step()
 
         mock_data = MagicMock(name="RheoData")
         mock_data.metadata = {"test_mode": "relaxation"}
